src/requirements.txt
advertools
python-whois
argparse
streamlit